% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Process the raw data, deleting the useless data.}
\usage{
filter_data(unfilter_data, threshold = 10000)
}
\arguments{
\item{rawdata_dir}{,the dir of raw data.}
}
\value{
: the processed data.
}
\description{
Process the raw data, deleting the useless data.
}
\examples{
library(IdenHerb)
Pre_processdata(rawdata_dir = example)
}
