% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre_processdata.R
\name{Pre_processdata}
\alias{Pre_processdata}
\title{Process the raw data, deleting the useless data.}
\usage{
Pre_processdata(rawdata_dir)
}
\arguments{
\item{rawdata_dir}{,the dir of raw data.}
}
\value{
: the processed data.
}
\description{
Process the raw data, deleting the useless data.
}
\examples{
library(IdenHerb)
Pre_processdata(rawdata_dir = example)
}
